// src/core/types/userData.ts
export interface UserDataType {
    user_id: string;
    registration_date: string;
    brand: string;
    tracking_code: string;
    afp: string;
    language: string;
    type: string;
    size: string;
    name: string;
    nci: number;
    status: string;
    qualification_date: string;
    country: string;
    volume: number;
    first_deposit: number;
    first_deposit_date: string;
    withdrawals: number;
    net_deposits: number;
    generic1: string;
    customer_name: string;
    commission: number;
  }
  